正式更新破滅國境:scripts
git-svn-id: http://www.l2jtw.com:3690/svn/datapack@337 9bd44e3e-6552-0410-907e-a6ca40b856b6

+++ L2JTW_Datapack_CT2/data/scripts/handlers/admincommandhandlers/AdminEditNpc.java
@@ -488,7 +488,7 @@ public class AdminEditNpc implements IAdminCommandHandler
 		activeChar.sendPacket(adminReply);
 	}
 	
-	private void storeTradeList(int itemID, int price, int tradeListID, int order)
+	private void storeTradeList(int itemID, long price, int tradeListID, int order)
 	{
 		java.sql.Connection con = null;
 		try
@@ -515,7 +515,7 @@ public class AdminEditNpc implements IAdminCommandHandler
 		}
 	}
 	
-	private void updateTradeList(int itemID, int price, int tradeListID, int order)
+	private void updateTradeList(int itemID, long price, int tradeListID, int order)
 	{
 		java.sql.Connection con = null;
 		try
@@ -569,7 +569,7 @@ public class AdminEditNpc implements IAdminCommandHandler
 		}
 	}
 	
-	private int findOrderTradeList(int itemID, int price, int tradeListID)
+	private int findOrderTradeList(int itemID, long price, int tradeListID)
 	{
 		java.sql.Connection con = null;
 		int order = 0;
